Draft Schubert cells; Additions to Category theory appendix
New section on the decomposition of Grassmannians into Schubert cells. In Category theory appendix added basic examples of how to construct new categories from old e.g slice category and functor category (due to used terminology in the text) and basic examples of functors that are generally good to know. Added a proof that equivalences of categories are characterized by functors that are full+faithful+essentially surjective on objects (again because the fact is used in the proof that the category of vector bundles is equivalent to the category of f.g locally free O-module sheaves). [Latex has not been compiled from github]