principal G-bundles and Cech cohomology
Moved the definition of principal G-bundles to the section on cocycles
and rephrased a number of theorem in sections 2 and 3 in terms of
principal G-bundles instead of vector bundles. Got rid of some fussing
about vector bundles of inhomogeneous dimension. Added some results on
Cech cohomology but Chapter 4 isn’t complete yet. In particular,
homotopy invariance is missing.